+++ exec.go
@@ -29,12 +29,12 @@ func handleLua(input string) (string, uint8, error) {
 	chunk, err := l.CompileAndLoadLuaChunk("", []byte(cmdString), rt.TableValue(l.GlobalEnv()))
 	if err != nil && noexecute {
 		fmt.Println(err)
-	/*	if lerr, ok := err.(*lua.ApiError); ok {
-			if perr, ok := lerr.Cause.(*parse.Error); ok {
-				print(perr.Pos.Line == parse.EOF)
+		/*	if lerr, ok := err.(*lua.ApiError); ok {
+				if perr, ok := lerr.Cause.(*parse.Error); ok {
+					print(perr.Pos.Line == parse.EOF)
+				}
 			}
-		}
-	*/
+		*/
 		return cmdString, 125, err
 	}
 	// And if there's no syntax errors and -n isnt provided, run